+++ CHANGELOG.md
@@ -18,6 +18,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Improved str variants view : sorting by locus, grouped by allele.
 - Delivery report PDF export
 - New mosaicism tag option
+- Add or modify individuals' age or tissue type from case page
 
 ### Fixed
 - Fixed update OMIM command bug due to change in the header of the genemap2 file